refactor: 폴더구조 변경(util 이동)
