+++ pages/add.tsx
@@ -7,7 +7,7 @@ import ComboBox from '@/features/common/components/ComboBox'
 import DatePickerField from '@/features/common/components/DatepickerField'
 import Input from '@/features/common/components/Input'
 import Navbar from '@/features/common/components/Navbar'
-import { classNames } from '@/features/common/utils/classNames'
+import { classNames } from '@/utils/classNames'
 
 const NotoSans = Noto_Sans({
   weight: ['500', '400'],
