+++ pages/edit-state.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { useForm } from 'react-hook-form'
 
 import Navbar from '@/features/common/components/Navbar'
-import { classNames } from '@/features/common/utils/classNames'
+import { classNames } from '@/utils/classNames'
 
 const NotoSans = Noto_Sans({
   weight: ['500', '400'],
